Extract target into own file
